feat: link zustand with ProjectMain
